tests(all): refactor tests folding
