+++ packages/bridge/tests/events/on.ts
@@ -1,4 +1,4 @@
-import { expect, it, vi, afterEach, describe, beforeEach } from 'vitest';
+import { expect, it, vi, afterEach, beforeEach } from 'vitest';
 
 import { on } from '../../src/index.js';
 import { createWindow, type WindowSpy } from '../../test-utils/createWindow.js';
@@ -14,42 +14,36 @@ afterEach(() => {
   windowSpy.mockRestore();
 });
 
-describe('events', () => {
-  describe('on.ts', () => {
-    describe('on', () => {
-      it('should call listener in case, Telegram event was created', () => {
-        const listener = vi.fn();
-        on('viewport_changed', listener);
+it('should call listener in case, Telegram event was created', () => {
+  const listener = vi.fn();
+  on('viewport_changed', listener);
 
-        const eventData = {
-          height: 123,
-          width: 321,
-          is_expanded: false,
-          is_state_stable: false,
-        };
-        dispatchWindowMessageEvent('viewport_changed', eventData);
+  const eventData = {
+    height: 123,
+    width: 321,
+    is_expanded: false,
+    is_state_stable: false,
+  };
+  dispatchWindowMessageEvent('viewport_changed', eventData);
 
-        expect(listener).toHaveBeenCalledTimes(1);
-        expect(listener).toHaveBeenCalledWith(eventData);
-      });
+  expect(listener).toHaveBeenCalledTimes(1);
+  expect(listener).toHaveBeenCalledWith(eventData);
+});
 
-      it('should remove listener in case, returned callback was called', () => {
-        const listener = vi.fn();
-        const emit = () => dispatchWindowMessageEvent('viewport_changed', {
-          height: 123,
-          width: 321,
-          is_expanded: false,
-          is_state_stable: false,
-        });
-
-        const off = on('viewport_changed', listener);
-        emit();
-        expect(listener).toHaveBeenCalledTimes(1);
-
-        off();
-        emit();
-        expect(listener).toHaveBeenCalledTimes(1);
-      });
-    });
+it('should remove listener in case, returned callback was called', () => {
+  const listener = vi.fn();
+  const emit = () => dispatchWindowMessageEvent('viewport_changed', {
+    height: 123,
+    width: 321,
+    is_expanded: false,
+    is_state_stable: false,
   });
+
+  const off = on('viewport_changed', listener);
+  emit();
+  expect(listener).toHaveBeenCalledTimes(1);
+
+  off();
+  emit();
+  expect(listener).toHaveBeenCalledTimes(1);
 });
