+++ packages/bridge/tests/events/parsers/invoiceClosed.ts
@@ -1,0 +1,12 @@
+import { expect, it } from 'vitest';
+
+import { invoiceClosed } from '../../../src/index.js';
+
+it('should return parsed value in case, passed value satisfies schema', () => {
+  const value = { slug: 'abc', status: 'def' };
+  expect(invoiceClosed().parse(value)).toStrictEqual(value);
+});
+
+it('should throw an error in case, passed value does not satisfy schema', () => {
+  expect(() => invoiceClosed().parse({})).toThrow();
+});
