+++ packages/bridge/tests/events/parsers/qrTextReceived.ts
@@ -1,0 +1,9 @@
+import { expect, it } from 'vitest';
+
+import { qrTextReceived } from '../../../src/index.js';
+
+it('should return parsed value in case, passed value satisfies schema', () => {
+  expect(qrTextReceived().parse({ data: 'ok' })).toStrictEqual({ data: 'ok' });
+  expect(qrTextReceived().parse({})).toStrictEqual({});
+  expect(qrTextReceived().parse({ data: 100 })).toStrictEqual({ data: '100' });
+});
