+++ packages/bridge/tests/events/subscribe.ts
@@ -1,4 +1,4 @@
-import { expect, describe, it, vi, beforeEach, afterEach } from 'vitest';
+import { expect, it, vi, beforeEach, afterEach } from 'vitest';
 
 import { subscribe } from '../../src/index.js';
 import { createWindow, type WindowSpy } from '../../test-utils/createWindow.js';
@@ -14,42 +14,36 @@ afterEach(() => {
   windowSpy.mockRestore();
 });
 
-describe('events', () => {
-  describe('subscribe.ts', () => {
-    describe('subscribe', () => {
-     it('should call listener in case, Telegram event was created', () => {
-        const listener = vi.fn();
-        subscribe(listener);
+it('should call listener in case, Telegram event was created', () => {
+  const listener = vi.fn();
+  subscribe(listener);
 
-        const eventData = {
-          height: 123,
-          width: 321,
-          is_expanded: false,
-          is_state_stable: false,
-        };
-        dispatchWindowMessageEvent('viewport_changed', eventData);
+  const eventData = {
+    height: 123,
+    width: 321,
+    is_expanded: false,
+    is_state_stable: false,
+  };
+  dispatchWindowMessageEvent('viewport_changed', eventData);
 
-        expect(listener).toHaveBeenCalledTimes(1);
-        expect(listener).toHaveBeenCalledWith('viewport_changed', eventData);
-      });
+  expect(listener).toHaveBeenCalledTimes(1);
+  expect(listener).toHaveBeenCalledWith('viewport_changed', eventData);
+});
 
-     it('should remove listener in case, returned callback was called', () => {
-        const listener = vi.fn();
-        const emit = () => dispatchWindowMessageEvent('viewport_changed', {
-          height: 123,
-          width: 321,
-          is_expanded: false,
-          is_state_stable: false,
-        });
-
-        const unsubscribe = subscribe(listener);
-        emit();
-        expect(listener).toHaveBeenCalledTimes(1);
-
-        unsubscribe();
-        emit();
-        expect(listener).toHaveBeenCalledTimes(1);
-      });
-    });
+it('should remove listener in case, returned callback was called', () => {
+  const listener = vi.fn();
+  const emit = () => dispatchWindowMessageEvent('viewport_changed', {
+    height: 123,
+    width: 321,
+    is_expanded: false,
+    is_state_stable: false,
   });
+
+  const unsubscribe = subscribe(listener);
+  emit();
+  expect(listener).toHaveBeenCalledTimes(1);
+
+  unsubscribe();
+  emit();
+  expect(listener).toHaveBeenCalledTimes(1);
 });
