+++ packages/bridge/tests/events/unsubscribe.ts
@@ -1,4 +1,4 @@
-import { expect, it, vi, beforeEach, afterEach, describe } from 'vitest';
+import { expect, it, vi, beforeEach, afterEach } from 'vitest';
 
 import { subscribe, unsubscribe } from '../../src/index.js';
 import { createWindow, type WindowSpy } from '../../test-utils/createWindow.js';
@@ -14,26 +14,20 @@ afterEach(() => {
   windowSpy.mockRestore();
 });
 
-describe('events', () => {
-  describe('unsubscribe.ts', () => {
-    describe('unsubscribe', () => {
-     it('should remove listener', () => {
-        const listener = vi.fn();
-        const emit = () => dispatchWindowMessageEvent('viewport_changed', {
-          height: 123,
-          width: 321,
-          is_expanded: false,
-          is_state_stable: false,
-        });
-
-        subscribe(listener);
-        emit();
-        expect(listener).toHaveBeenCalledTimes(1);
-
-        unsubscribe(listener);
-        emit();
-        expect(listener).toHaveBeenCalledTimes(1);
-      });
-    });
+it('should remove listener', () => {
+  const listener = vi.fn();
+  const emit = () => dispatchWindowMessageEvent('viewport_changed', {
+    height: 123,
+    width: 321,
+    is_expanded: false,
+    is_state_stable: false,
   });
+
+  subscribe(listener);
+  emit();
+  expect(listener).toHaveBeenCalledTimes(1);
+
+  unsubscribe(listener);
+  emit();
+  expect(listener).toHaveBeenCalledTimes(1);
 });
